Control scripts (like config but a bit more powerful) for cosmic.
Can interact with MASS via moo, and Lotus via bsub.
Depends on cosmic package for cosmic_ cmds and python utils.
These are launched by e.g.:

* cosmic_retrieve_from_mass u-al508_ap9_precip.py
* cosmic_bsub_submit cmorph_extract_asia_8km_30min_ctrl.py

The control script's job is to provide the necessary information for a
run script. A run script will end up being called with:
python <run_script> <ctrl_script> <key>

where <key> is a string key that tells that instance which of the tasks
it is to perform.

Split into different directories by usage.

* MASS: get UM data from MASS.
* UM: do standard conversion/extraction of Asia from UM data.
* cmorph: download, convert and extract CMORPH data.
* WP2_analysis: perform analysis for WP2 of COSMIC.
* experimental: try out new ideas.